requirements.txt
streamlit
pypdf
google-genai
python-dotenv

